Split into image reader class and tiff converter utility files
